Matrix Inversion and Caching the Result
Calculates the Inversion of the special matrix created. It first checks to see if the Inverse has already been calculated. If so, it gets the Inverse from the cache and skips the computation. Otherwise, it calculates the Inverse of the Matrix and sets the value of the in the cache.